requirements.txt
pillow>=10.4.0
numpy>=1.26.0
pytest>=8.0.0
